==> 1.human_muring_GSEA.Rmd <==
---
title: "Human and murine responses to Mtb"
subtitle: "GSEA"
author: "Kim Dill-McFarland, kadm@uw.edu"
output:
  html_document:
    toc: yes
    toc_float:
      collapsed: no
date: "version `r format(Sys.time(), '%B %d, %Y')`"
editor_options: 
  chunk_output_type: console
---
# Setup
Load packages

```{r message=FALSE, warning=FALSE}
# Data manipulation and figures
library(tidyverse)
library(BIGpicture)
library(patchwork)
# library(readxl)
#Print pretty tables to Rmd
library(DT)
```

Set seed

```{r}
set.seed(589)
```

# Load data

```{r data}
H_hum <- read_csv("data_raw/human/AM-MDM/h_GSEA.result.csv") 

H_mur <- read_csv("data_raw/murine/GSEA_AM_H37Rv_RNAseq.csv")
```

# GSEA term overlap

```{r echo=FALSE}
H_hum_format <- H_hum %>% 
  filter(group %in% c("AM","MDM")) %>% 
  mutate(group = recode(group,
                        "AM"="Human AM",
                        "MDM"="Human MDM")) %>% 
  select(group, pathway, fgsea.FDR, fgsea.NES) %>% 
  rename(FDR=fgsea.FDR, NES=fgsea.NES) %>% 
  mutate(pathway = gsub("HALLMARK_","",pathway))

H_mur_format <- H_mur %>% 
  mutate(group = recode(Sample,
                        "CoTB_Mtb"="Murine AM coTB",
                        "scBCG_Mtb"="Murine AM scBCG",
                        "Unvax_Mtb"="Murine AM")) %>% 
  select(group, Pathway, `FDR q-val`, NES) %>%
  rename(FDR="FDR q-val", pathway=Pathway)

# check pathway names
table(unique(H_hum_format$pathway) %in% H_mur_format$pathway)

#combine
H_all <- bind_rows(H_hum_format, H_mur_format)
# check no addtl pathways
length(unique(H_all$pathway))
```

```{r echo=FALSe}
#list all same direction 
H_AM_wide <- H_all %>% 
  filter(grepl("AM", group)) %>% 
  select(-FDR) %>% 
  pivot_wider(names_from = group, values_from = NES)

shared_AM <- H_AM_wide[(!rowSums(H_AM_wide[c(-1,-3)] > 0)) | (!rowSums(H_AM_wide[-1] < 0)), ]

plotH1 <- H_all %>% 
  filter(grepl("AM", group) & pathway %in% shared_AM$pathway) %>% 
  mutate(group = paste(group,"Down with Mtb <--    --> Up with Mtb", 
                       sep="\n")) %>% 
plot_gsea(., fdr.cutoff = 1.1) +
  ggtitle("Shared human and murine AM responses to Mtb infection")
```

```{r echo=FALSE}
#different directions
plotH2 <- H_all %>% 
  filter(grepl("AM", group) & !pathway %in% shared_AM$pathway) %>% 
  mutate(group = paste(group,"Down with Mtb <--    --> Up with Mtb", 
                       sep="\n")) %>% 
plot_gsea(., fdr.cutoff = 1.1) +
  ggtitle("Unique human and murine AM responses to Mtb infection")
```

```{r}
plotH1/plotH2 +
  plot_layout(widths = c(nrow(shared_AM), 50-nrow(shared_AM)))
```

# R session

```{r}
sessionInfo()
```

***
